feat(gameplay): add xp gem drops
Spawn GemBlue pickups from enemy xp values, apply stage XP bonuses, and collect nearby gems into the player experience total.

+++ src/Entities/Enemy/EnemyBase.cpp
@@ -159,7 +159,7 @@ float EnemyBase::GetMass() const
     return m_stats.mass;
 }
 
-int EnemyBase::GetExpYield() const
+float EnemyBase::GetExpYield() const
 {
     return m_stats.expYield;
 }
